Restructure components into domain folders
