+++ __tests__/ThemeToggle.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
-import Layout from '../components/Layout';
+import Layout from '../components/Layout/Layout';
 import { AppContext } from '../contexts/AppContext';
 
 jest.mock('next/router', () => ({
