+++ components/Category/CategoryCard.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import Link from 'next/link';
 import Image from 'next/image';
-import type { Category } from '../types';
-import CATEGORY_IMAGES from '../lib/categoryImages';
+import type { Category } from '../../types';
+import CATEGORY_IMAGES from '../../lib/categoryImages';
 
 interface CategoryCardProps {
   category: Category;
