+++ components/Category/CategoryPromotion.tsx
