+++ components/Category/CategorySlider.tsx
