+++ components/Category/CreateCategoryModal.tsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from 'react';
-import { GenericInput, GenericModal } from './ui';
-import { slugify } from '../lib/slugify';
+import { GenericInput } from '../UI';
+import GenericModal from '../Modals/GenericModal';
+import { slugify } from '../../lib/slugify';
 
 interface CreateCategoryModalProps {
   isOpen: boolean;
