+++ components/Layout/Layout.tsx
@@ -1,5 +1,5 @@
 import React, { FC, ReactNode, Dispatch, SetStateAction } from 'react';
-import useTheme, { Theme } from '../hooks/useTheme';
+import useTheme, { Theme } from '../../hooks/useTheme';
 import Header from './Header';
 import Footer from './Footer';
