+++ components/Layout/UserHeader.tsx
@@ -4,22 +4,22 @@ import { useContext, useState, useEffect, useRef } from 'react';
 import { useRouter } from 'next/router';
 import { useSession, signOut } from 'next-auth/react';
 import type { FC } from 'react';
-import { AppContext } from '../contexts/AppContext';
-import CartIcon from './icons/CartIcon';
-import MoonIcon from './icons/MoonIcon';
-import SunIcon from './icons/SunIcon';
-import UserIcon from './icons/UserIcon';
-import MenuIcon from './icons/MenuIcon';
-import ChevronDownIcon from './icons/ChevronDownIcon';
-import ElectronicsIcon from './icons/ElectronicsIcon';
-import FashionIcon from './icons/FashionIcon';
-import HomeIcon from './icons/HomeIcon';
-import ToysIcon from './icons/ToysIcon';
-import SportsIcon from './icons/SportsIcon';
-import DEFAULT_CATEGORIES from '../lib/defaultCategories';
+import { AppContext } from '../../contexts/AppContext';
+import CartIcon from '../icons/CartIcon';
+import MoonIcon from '../icons/MoonIcon';
+import SunIcon from '../icons/SunIcon';
+import UserIcon from '../icons/UserIcon';
+import MenuIcon from '../icons/MenuIcon';
+import ChevronDownIcon from '../icons/ChevronDownIcon';
+import ElectronicsIcon from '../icons/ElectronicsIcon';
+import FashionIcon from '../icons/FashionIcon';
+import HomeIcon from '../icons/HomeIcon';
+import ToysIcon from '../icons/ToysIcon';
+import SportsIcon from '../icons/SportsIcon';
+import DEFAULT_CATEGORIES from '../../lib/defaultCategories';
 import SearchBar from './SearchBar';
-import type { Category } from '../types/category';
-import type { User } from '../types/user';
+import type { Category } from '../../types/category';
+import type { User } from '../../types/user';
 
 interface HeaderProps {
   theme?: string;
