+++ components/Product/ProductGrid.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ProductCard from './ProductCard';
 import ProductCardSkeleton from './ProductCardSkeleton';
-import type { Product } from '../types/product';
+import type { Product } from '../../types/product';
 
 interface ProductGridProps {
   products: Product[];
