+++ components/Product/ProductImageSlider.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useRef, useEffect } from 'react';
 import Image from 'next/image';
-import type { Image as ProductImage } from '../types/image';
-import ChevronLeftIcon from './icons/ChevronLeftIcon';
-import ChevronRightIcon from './icons/ChevronRightIcon';
+import type { Image as ProductImage } from '../../types/image';
+import ChevronLeftIcon from '../icons/ChevronLeftIcon';
+import ChevronRightIcon from '../icons/ChevronRightIcon';
 
 export interface ProductImageSliderProps {
   images?: ProductImage[];
