+++ components/UI/index.ts
@@ -1,3 +1,3 @@
 export { default as InputField } from './InputField';
 export { default as GenericInput } from './GenericInput';
-export { default as GenericModal } from './GenericModal';
+export { default as GenericModal } from '../Modals/GenericModal';
