+++ pages/categories/[slug]/products.tsx
@@ -3,7 +3,7 @@ import Head from 'next/head';
 import { getPageTitle } from '../../../lib/pageTitle';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
-import ProductCard from '../../../components/ProductCard';
+import ProductCard from '../../../components/Product/ProductCard';
 import type { Product, Category } from '../../../types';
 import {
   getCategoryBySlug,
