+++ pages/index.tsx
@@ -1,10 +1,10 @@
 import Head from 'next/head';
 import { useEffect, useState } from 'react';
 import HomeHero from '../components/HomeHero';
-import FeaturedProducts from '../components/FeaturedProducts';
-import CategoryPromotion from '../components/CategoryPromotion';
+import FeaturedProducts from '../components/Product/FeaturedProducts';
+import CategoryPromotion from '../components/Category/CategoryPromotion';
 import PromoBanner from '../components/PromoBanner';
-import CategorySlider, { CategoryItem } from '../components/CategorySlider';
+import CategorySlider, { CategoryItem } from '../components/Category/CategorySlider';
 import { getPageTitle } from '../lib/pageTitle';
 import DEFAULT_CATEGORIES from '../lib/defaultCategories';
 import CATEGORY_IMAGES from '../lib/categoryImages';
