+++ pages/products/index.tsx
@@ -2,9 +2,9 @@ import { GetServerSideProps } from 'next';
 import Head from 'next/head';
 import { useRouter } from 'next/router';
 import { useCallback, useEffect, useRef, useState } from 'react';
-import ProductFilters from '../../components/ProductFilters';
+import ProductFilters from '../../components/Product/ProductFilters';
 import ActiveFilters from '../../components/ActiveFilters';
-import ProductGrid from '../../components/ProductGrid';
+import ProductGrid from '../../components/Product/ProductGrid';
 import Loader from '../../components/Loader';
 import InfiniteLoader from '../../components/InfiniteLoader';
 import SortMenu, { SortValue } from '../../components/SortMenu';
