+++ pages/search.tsx
@@ -10,7 +10,7 @@ import Head from 'next/head';
 import Link from 'next/link';
 import { getPageTitle } from '../lib/pageTitle';
 import HeroSlider from '../components/HeroSlider';
-import ProductCard from '../components/ProductCard';
+import ProductCard from '../components/Product/ProductCard';
 import DEFAULT_CATEGORIES from '../lib/defaultCategories';
 import { Product } from '../types/product';
